Renamed router files and changed path in index.js to conform to new folder-structure
