+++ index.js
@@ -1,8 +1,8 @@
 import { initializeApp } from "firebase/app"
 import express from 'express'
 
-import positionRouter from "./routes/position.mjs"
-import imageRouter from "./routes/image.mjs"
+import positionRouter from "./presentation/positionRouter.mjs"
+import imageRouter from "./presentation/imageRouter.mjs"
 
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
